Update simpleTkGUI.py
This script creates a tkinter GUI with two switch buttons and an exit button. 
The GPIO outputs correctly illuminate an LED on pins 13 & 15, NOT GPIOs 13 & 15.
The capacitive screen correctly presses these buttons. In regard to cursor/button presses, note that screen orientation can mismatch with the hardware orientation when the screen is rotated using the adafruit-pitft.py setup script. 

- Use extreme care when reading the Raspberry Pin pinout and carefully compare it with labels on any raspberry pi-breakouts, cobblers, or black-hat breakouts. The silkscreened pin labels on the older raspberry pi cobbler read "27" for pin 13, and "22" for pin 15. On the Pi400 these correspond to GPIOs 3 & 5.

Next Steps:
1. Find out how to close the tkinter window and open a new one when changing states
2. Determine the x,y positions of a rotated GUI, find out how to correct this for any screen orientation