Update ltbF22int_Rev0.py
The tkinter GUI exists and functions in the Home State. Debugging is required to transition to the next state. Debugging the GPIO has characterized the main challenges in integrating these parts. The next step will be to integrate the same GUI to all states and change the labels to the appropriate functions/features.