Update and rename ltbF22int_Rev1.py to ltbF22int_Rev2.py
The State Machine changes state from the Home State after pressing a GUI button. The GUI button causes the window to close after 5 seconds and the machine changes state to Profile1 or Profile2. Next steps is to copy and paste the GUI into all states and verify operation. Noe that when hooking up the circuit, the input pins refer to GPIOs and Outputs refer to the hardware pin numbers of the GPIO s. Additionally, the gui_btn1 and gui_btn2 variables are intended to store a boolean that will be an internal software version of the hardware output.