Add files via upload
Integrated state machine and tkinter GUI, complete with timestamps. This script requires a folder to be placed on the desktop of the Raspberry Pi named, "ltb_f22_release". The script should create the .csv files named, "P1_t1.csv", "P1_t2.csv", "P2_t1.csv", and "P2_t2.csv". The file names represent all of the time logged for each task. For example P1_t1 is Profile 1, task 1. This code needs a coded "Focus Timer" and a count up timer that appears in the window of each task when actively tracking that task.